[Refactor] Separate Google login logic
FirebaseManager와 구글 로그인 기능 분리
