+++ Assets/07.CYH_Folder/Scripts/GoogleLogin.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5856fb1f6a722104bacf79bb604fef43
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
