+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,21 +6,6 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/04.Scenes/LoginScene.unity
-    guid: 991a7c76d093752479f56b41ff11c4ef
-  - enabled: 1
-    path: Assets/04.Scenes/MainScene.unity
-    guid: 028fbc514aa4d23499bfa0b822a87db5
-  - enabled: 1
-    path: Assets/04.Scenes/InGameScene/GameScene_Map1.unity
-    guid: ed802662982d1514d995c49f2fcf01e1
-  - enabled: 1
-    path: Assets/04.Scenes/InGameScene/GameScene_Map2.unity
-    guid: 4ab47b1c58c3a464a837d7d1475da3a7
-  - enabled: 1
-    path: Assets/04.Scenes/InGameScene/GameScene_Map3.unity
-    guid: d9a3478125a2d98479e606395886a172
-  - enabled: 1
-    path: Assets/04.Scenes/InGameScene/GameScene_Map4.unity
-    guid: d381b6c1ad930ef43ab8178ea3ac4b7e
+    path: Assets/07.CYH_Folder/Scenes/[CYH] LoginScene.unity
+    guid: 66482cb27c8936546ab09a9672037326
   m_configObjects: {}
